// js-client/src/client/models/Page_CategoryOut_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { CategoryOut } from './CategoryOut';

export type Page_CategoryOut_ = {
    items: Array<CategoryOut>;
    total?: number;
    page?: number;
    size?: number;
    pages?: number;
};