// js-client/src/client/models/Page_LikeOut_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { LikeOut } from './LikeOut';

export type Page_LikeOut_ = {
    items: Array<LikeOut>;
    total?: number;
    page?: number;
    size?: number;
    pages?: number;
};