// js-client/src/client/models/Page_PostOut_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { PostOut } from './PostOut';

export type Page_PostOut_ = {
    items: Array<PostOut>;
    total?: number;
    page?: number;
    size?: number;
    pages?: number;
};