// js-client/src/client/models/Page_ThreadOut_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ThreadOut } from './ThreadOut';

export type Page_ThreadOut_ = {
    items: Array<ThreadOut>;
    total?: number;
    page?: number;
    size?: number;
    pages?: number;
};